move blueoil/nn/layers to blueoil/layers
