+++ lmnet/lmnet/blocks.py
@@ -15,7 +15,7 @@
 # =============================================================================
 import tensorflow as tf
 
-from blueoil.nn.layers import batch_norm, conv2d
+from blueoil.layers import batch_norm, conv2d
 
 
 # TODO(wakisaka): should be replace to conv_bn_act().
