+++ lmnet/lmnet/networks/classification/darknet.py
@@ -18,7 +18,7 @@ from functools import partial
 import tensorflow as tf
 
 from lmnet.blocks import darknet as darknet_block
-from blueoil.nn.layers import conv2d, max_pooling2d
+from blueoil.layers import conv2d, max_pooling2d
 from lmnet.networks.classification.base import Base
