+++ lmnet/lmnet/networks/classification/resnet.py
@@ -15,7 +15,7 @@
 # =============================================================================
 import tensorflow as tf
 
-from blueoil.nn.layers import average_pooling2d, batch_norm, conv2d, fully_connected
+from blueoil.layers import average_pooling2d, batch_norm, conv2d, fully_connected
 from lmnet.networks.classification.base import Base
