+++ lmnet/lmnet/networks/object_detection/lm_fyolo.py
@@ -18,7 +18,7 @@ from functools import partial
 import tensorflow as tf
 
 from lmnet.blocks import darknet
-from blueoil.nn.layers import conv2d
+from blueoil.layers import conv2d
 from lmnet.networks.object_detection.yolo_v2 import YoloV2
