+++ lmnet/lmnet/networks/object_detection/yolo_v1.py
@@ -16,7 +16,7 @@
 import numpy as np
 import tensorflow as tf
 
-from blueoil.nn.layers import conv2d, fully_connected, max_pooling2d
+from blueoil.layers import conv2d, fully_connected, max_pooling2d
 from lmnet.metrics.mean_average_precision import average_precision, tp_fp_in_the_image
 from lmnet.networks.base import BaseNetwork
