+++ lmnet/lmnet/networks/object_detection/yolo_v2.py
@@ -19,7 +19,7 @@ import numpy as np
 import tensorflow as tf
 
 from lmnet.blocks import darknet as darknet_block
-from blueoil.nn.layers import conv2d, max_pooling2d
+from blueoil.layers import conv2d, max_pooling2d
 from lmnet.metrics.mean_average_precision import average_precision, tp_fp_in_the_image
 from lmnet.networks.base import BaseNetwork
