+++ lmnet/tests/lmnet_tests/test_layers.py
@@ -16,8 +16,8 @@
 import numpy as np
 import tensorflow as tf
 
-from blueoil.nn.layers import max_pool_with_argmax
-from blueoil.nn.layers.experiment import max_unpool_with_argmax
+from blueoil.layers import max_pool_with_argmax
+from blueoil.layers.experiment import max_unpool_with_argmax
 
 
 def test_unpool():
